fix: Simplify LIST output handling using Value::LIST() to fix BatchedDataCollection error
- Replace low-level ListVector manipulation with Value::LIST() approach
- Remove custom GlobalState that forced single-threaded execution
- Restore ORDER BY in macro for correct BOCPD sequential processing

This fixes the BatchedDataCollection::Merge error that occurred when
running ts_detect_changepoints_by without COUNT(*) wrapper. Memory
usage is now similar to the old LIST-based implementation (~48%
improvement at 1M rows due to simpler output handling).

+++ src/macros/ts_macros.cpp
@@ -451,7 +451,7 @@ ORDER BY od._dt
     {"ts_detect_changepoints_by", {"source", "group_col", "date_col", "value_col", "params", nullptr}, {{nullptr, nullptr}},
 R"(
 SELECT * FROM _ts_detect_changepoints_native(
-    (SELECT group_col, date_col, value_col FROM query_table(source::VARCHAR) ORDER BY group_col, date_col),
+    (SELECT group_col, date_col, value_col::DOUBLE FROM query_table(source::VARCHAR) ORDER BY group_col, date_col),
     COALESCE(TRY_CAST(json_extract_string(to_json(params), '$.hazard_lambda') AS DOUBLE), 250.0),
     COALESCE(TRY_CAST(json_extract_string(to_json(params), '$.include_probabilities') AS BOOLEAN), false)
 )
